refactor: create getMultisigForNetwork util and reuse it across components
